requirements.txt
chromadb
google-generativeai
langchain
langchain
langchain_experimental
langchain_google_genai
openai
pathlib
python-dotenv
streamlit
tabulate
sqlalchemy